Includes Sign In page and Home page
