+++ GUI/Screens/__init__.py
@@ -1,0 +1,2 @@
+from .HomePage import HomePage
+from .SignInPage import SignInPage
